Move multi component loader test to separate file
