+++ internal/packages/packageloader/loader_test.go
@@ -135,28 +135,6 @@ func TestLoader(t *testing.T) {
 	}, spec.Phases)
 }
 
-func TestMultiComponentLoader(t *testing.T) {
-	t.Parallel()
-
-	transformer, err := packageloader.NewTemplateTransformer(
-		packageloader.PackageFileTemplateContext{
-			Package: manifestsv1alpha1.TemplateContextPackage{
-				TemplateContextObjectMeta: manifestsv1alpha1.TemplateContextObjectMeta{Namespace: "test123-ns"},
-			},
-		},
-	)
-	require.NoError(t, err)
-
-	l := packageloader.New(testScheme, packageloader.WithDefaults, packageloader.WithFilesTransformers(transformer))
-
-	ctx := logr.NewContext(context.Background(), testr.New(t))
-	files, err := packageimport.Folder(ctx, filepath.Join("testdata", "multi-component", "000-ok-components-disabled"))
-	require.NoError(t, err)
-
-	_, err = l.FromFiles(ctx, files)
-	require.NoError(t, err)
-}
-
 var testPackageManifestContent = `apiVersion: manifests.package-operator.run/v1alpha1
 kind: PackageManifest
 metadata:
