+++ internal/packages/packageloader/testdata/multi-component/000-ok-components-disabled/manifest.yaml
@@ -9,7 +9,6 @@ spec:
     - name: configure
     - name: deploy-backend
     - name: deploy-frontend
-    - name: deploy-monitoring
   availabilityProbes:
     - probes:
         - condition:
